Move hash utility functions to a new file
